backend-ocr/requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
easyocr>=1.7.0
pillow>=11.0.0
python-multipart>=0.0.6
